parts/apicheck.pl: Don't skip some cases
We only should create a single test for a given function.  But if that
function is like this:

  #if foo
    int bar(...)
  #else
    char bar(...)
  #endif

We would only generate a test for the 'foo' case.  Uniqueness of a
function includes the compile time conditional expression for it.

(cherry picked from commit 2d6698cc1724195e18acd67fb8fddf9527372eda)
Signed-off-by: Nicolas R <atoomic@cpan.org>