parts/inc/utf8: Symbols are not really for external use
This used __UNDEFINED__ to define a bunch of constants.  However, there
is nothing backported that allows the user to specify any of these, so
there is no need for them to be publicly accessible.  And, if something
is important enough to be backported, it ought to be documented.  These
symbols should not be documented, so __UNDEFINED__ is not the proper way
to deal with them.

(cherry picked from commit 359209af4e95548cbd7291128549ce2071e8e555)
Signed-off-by: Nicolas R <atoomic@cpan.org>