parts/ppptools.pl: Make sure mathoms fcns are deprecated
This makes sure that we consider functions in mathoms as being
deprecated.  The current blead regen/embed.pl already does this, but
this makes sure, should that get reverted.  A mathoms function with
an M flag indicates there is a macro by that name that avoids the
function (which must begin with 'Perl_'), and the macro isn't deprecated
unless the D flag is also present.

(cherry picked from commit 8ed3f9e76aca7a9341a8878d2d7c61a8b8717621)
Signed-off-by: Nicolas R <atoomic@cpan.org>