PPPort_pm.PL: Add an item for unknown min supported version
The pod for this has sections of what elements are first supported when.
Unfortunately there are some that we do not have that information.

The cure for this is for someone to contribute pod for that element to
the perl core, but in the meantime, this lists those we couldn't figure
out.

It uses ppport.fnc to figure out many of the ones it otherwise couldn't.
And it is possible to enhance this to figure out more.  But just
documenting them removes them from the list.  And they should be
documented anyway.

(cherry picked from commit 5b548c97ec59bfb65f3caefea9fcf50cf8cd6d49)
Signed-off-by: Nicolas R <atoomic@cpan.org>