Pass values from mktodo.pl to apicheck.pl
This uses an environment variable to send the version number across a
make to apicheck.pl, where it is inserted as comments in apicheck.c.

I found that when debugging, it is useful to have this information
handy throughout the file.

(cherry picked from commit 2a055a1a8526c1b721d10078ce9bf39178c7ea62)
Signed-off-by: Nicolas R <atoomic@cpan.org>