Handle "literal string" args
Some of the macros in perl require an argument to be a literal C string,
that is, one enclosed in quotation marks.  This change causes trim_arg()
to recognize these and return them properly, and apicheck.pl to use the
passed string directly as the argument.

(cherry picked from commit bc85e4126fa938abcaecadef4cbd8678da215250)
Signed-off-by: Nicolas R <atoomic@cpan.org>