embed.fnc: Use M flag generally with b, not m.
Commit 54c193aee0eb1f20806952d8fcd2ddadd5e57600 introduced the M flag,
to indicate an extra macro in addition to a regular function, beyond the
m flag which indicates a macro that replaces the function.  This
distinction matters to downstream processes.

But some of the functions with the b flag (hence in mathoms) still had
an m flag, but the macro is actually extra.  This commit changes those
from m to M.