Prepare VMS linker GSMATCH for 5.32.
GSMATCH in the linker options file tells the image activator
whether a dynamic library being loaded is compatible with what it
originally linked against.  Ideally we would like to encode the
version number and all the options that can make a build binary
incompatible with another build of the same version, but there
are about 30 such options and only 8 bits (on Alpha) in which to
store both the version number and the options.

We've lived with only 3 bits reserved for 3 options, but as of
version 32 of Perl 5, we'll need 6 bits for the version number:

    $ perl -e "$v = 31; printf('%b', $v);"
    11111
    $ perl -e "$v = 32; printf('%b', $v);"
    100000

leaving us only 2 bits for options.

The code modified here only kicks in when PERLSHR_USE_GSMATCH is
defined in the environment.