Fix parent directory in pod2usage auxiliary script.
File::Basename::dirname does not always remove the last component
of the path if that component is already a directory, depending on
platform.  This behavior is documented. pod2usage.t has been
failing on VMS because we get the wrong directory into @INC.

So, since we already have File::Spec loaded, just use the more
obvious and portable method for getting the parent directory.

Reported upstream at:

   <https://rt.cpan.org/Ticket/Display.html?id=130418>