PATCH: [perl #134405] Compiler warnings in blead
This was caused by functions that should have been deprecated all along,
and now are.  But they are still called in places.  In most instances
one deprecated function is a wrapper for another.  In one instance, the
calling of the function is to just make sure that mathoms.o gets linked
to.

The solution adopted here, some of which was suggested by Tony Cook, is
to use #pragmas to silence the warnings.  By doing this around the
entirety of mathoms.c, future issues will most likely be avoided.  And
there are unlikely to enough future cases outside of mathoms to cause a
problem.