Add ANYOFR regnode
This matches a single range of code points.  It is both faster and
smaller than other ANYOF-type nodes, requiring, after set-up, a single
subtraction and conditional branch.

The vast majority of Unicode properties match a single range, though
most of these are not likely to be used in real world applications.  But
things like [ij] are a single range, and those are quite commonly
encountered.  This matches them more efficiently than a bitmap would,
and doesn't require the space for one either.

The flags field is used to store the minimum matchable start byte for
UTF-8 strings, and is ignored for non-UTF-8 targets.  This, like ANYOFH
nodes which have the same mechanism, allows for quick weeding out of
many possible matches without having to convert the UTF-8 to its
corresponding code point.

This regnode packs the 32 bit argument with 20 bits for the minimum code
point the node matches, and 12 bits for the maximum range.  Values
outside those simply won't compile to this regnode, instead going to one
of the ANYOFH flavors.  This is sufficient to match all of Unicode
except for the final (private use) 65K plane.