regcomp.sym: Move regnodes to end that don't use next_off
Most regnodes use the next_off field in a regnode structure, to link to
the next one in the chain.  But some require more than the 16 bits it
contains, so they use a different, 32 bit, field.

Currently, there is a lookup array to distinguish between the types, but
that becomes unnecessary if all of one sort are grouped before or after
all of the other.