makedef.pl: export symbols flagged as C
0923255565af0741 broke the build on any platform that does not
promiscuously export all symbols from a shared library because
it moved a bunch of symbols from being API (A flag) to core-only
(C flag), but the new C flag was unknown to the program that
generates the export list.  Some of these functions are already
used outside the core (such as by the Encode extension) so we
can't simply pull the plug on them.