regcomp.c: Change handling of filled EXACT nodes
This changes the detection mechanism to, just before we otherwise would
write to the node, we see if that would be out of bounds, and if so,
instead break out of the loop to handle a full node.

This improves the packing of nodes, especially under /i, from the
previous mechanism.  But more importantly, it set things up so that we
can potentially increase the node size as we go along.

This also changes the handling of XXX