regcomp.c: Avoid reparsing a node
When it's determined that the final character of a node really should
be in the next node to avoid splitting a multi-character fold, the node
is refilled from scratch, setting an upper limit of how far to go just
before that character.

This commit changes to avoid that reparsing and refilling if just the
final character has to be omitted.  If the node has to be backed up two
or more characters, the reparsing is still done.  The technique in this
commit could be extended to avoid those, but doing it generally is
complicated, and doing it the way this commit does gets more and more
cumbersome, and the odds rapidly get smaller of needing to do more than
what I've done here.