Do not use 'exists' on arrays; use 'defined' instead
pod/perlfunc.pod advises us:

       WARNING: Calling "exists" on array values is strongly
       discouraged. The notion of deleting or checking the existence of
       Perl array elements is not conceptually coherent, and can lead
       to surprising behavior.

While trying to extend the test coverage of Dumpvalue.pm, two instances
of 'exists ARRAYREF' were noted.  In the second of these two instances,
using 'exists ARRAYREF' resulted in the 'if' condition of an if/else
block always returning true, thereby making the corresponding 'else'
condition unreachable -- a situation which was confirmed by coverage
analysis.

We should only use best practices in modules maintained by Perl 5
Porters; this patch implements that.  When we do so, a different (but
actually better) error message appears in the second instance; so tests
are updated as well.

For: RT 134492