Don't test for 16-bit inputs in inRANGE()
This macro has been expanding to overflow some assertion strings on
Windows and HP.  This commit omits handling one unlikely scenario,
namely that the input is a short, 16-bits.  And that is enough to get it
to compile on Windows.  HP isn't smoked on branches, so I don't know if
this will fix it.

More discussion may be needed, but this is all I have time for at the
moment.