make Scalar::Util::blessed() and pureperl Data::Dumper handle unicode class names properly
This also silences a warning from implicit-fallthrough in ListUtil.xs

This patch is annoying as we have to fix a module that is in the 'cpan'
namespace (Scalar::Util) in order to fix a module that is in the 'dist'
namespace (Data::Dumper), which IMO is not a good state of affairs. We
should not need to depend on a cpan module being changed to fix core.