Speedup Travis CI: stop smoking on Darwin
Resolves #17209

A full Travis CI build time is about 4 hours,
most of the time is spent on osx servers.

Travis CI does not seem to have the same kind
of resources for Darwin than it has for linux.

Disabling Darwin will speedup the builds to ~15 min

Also enable fast finish to stop on first failure.