XS-APItest: add tests for U8TO64_LE() and other hashing macros and code
Includes testing siphash24 and siphash13 properly, especially
testing against the SipHash 2-4 reference test vector,
see https://131002.net/siphash/siphash24.c

See also #17244 where we originally discovered there were no
tests for the internals of the hashing code and that because
of it we let slip in a very broken patch to the code.

Thanks to James E Keenan, Tony Cook and Hugo Van der Sanden
for support putting this together.

XS-APItest: fixup issues with # directives

for some reason the indentation of the #if clauses broke things
under some builds, but not all. Possibly a ccache issue.