Perl_perly_sighandler: combine two conditions
The branch is only taken if a number of conditions are true. Combine
them into a single if statement, with the sip test first, so that an
unnecessary call to sigaction() may be skipped if its only going to
immediately fail afterwards because sip isn't true.

Shouldn't be any change in functionality.