pat_advanced.t: Update test
This test was no longer exercising the code it once had, because ':'
doesn't go in a folded regnode.  Change to use a character that does
have a fold.

And doing so shows that something in the past broke this test.  This
branch will fix that; in the meantime make some tests TODO